+++ lib/src/inherited_model.dart
@@ -3,9 +3,10 @@ part of 'vxstate.dart';
 /// Tracks the listener widgets and notify them when
 /// their corresponding mutation executes
 class _VxStateModel extends InheritedModel {
-  final Set<Type> recent;
+  final Set<Type>? recent;
 
-  _VxStateModel({Widget child, this.recent}) : super(child: child);
+  const _VxStateModel({required Widget child, this.recent})
+      : super(child: child);
 
   @override
   bool updateShouldNotify(covariant InheritedWidget oldWidget) =>
@@ -16,6 +17,6 @@ class _VxStateModel extends InheritedModel {
       covariant InheritedModel oldWidget, Set dependencies) {
     // check if there is a mutation executed for which
     // dependent has listened
-    return dependencies.intersection(recent).isNotEmpty;
+    return dependencies.intersection(recent!).isNotEmpty;
   }
 }