+++ lib/src/widgets/vxbuilder.dart
@@ -4,19 +4,19 @@ import '../vxstate.dart';
 
 /// A stream builder like widget that accepts
 /// mutations and rebuilds after their execution.
-class VxBuilder extends StatelessWidget {
+class VxBuilder<T> extends StatelessWidget {
   /// [builder] provides the child widget to rendered.
-  final VxStateWidgetBuilder builder;
+  final VxStateWidgetBuilder<T> builder;
 
   /// Widget will rerender every time any of [mutations] executes.
   final Set<Type> mutations;
 
   /// Creates widget to rerender child widgets when given
   /// [mutations] execute.
-  VxBuilder({
-    @required this.builder,
-    @required this.mutations,
-  }) : assert(mutations != null);
+  const VxBuilder({
+    required this.builder,
+    required this.mutations,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -26,13 +26,13 @@ class VxBuilder extends StatelessWidget {
     return StreamBuilder<VxMutation>(
       stream: stream,
       builder: (context, mut) {
-        VxStatus status;
+        VxStatus? status;
         if (!mut.hasData || mut.connectionState == ConnectionState.waiting) {
           status = VxStatus.none;
         } else {
-          status = mut?.data?.status;
+          status = mut.data?.status;
         }
-        return builder(context, status);
+        return builder(context, VxState.store as T, status);
       },
     );
   }