+++ lib/src/widgets/vxconsumer.dart
@@ -7,42 +7,44 @@ import 'vxnotifier.dart';
 
 /// A stream builder like widget that accepts
 /// mutations and rebuilds after their execution.
-class VxConsumer extends StatefulWidget {
+class VxConsumer<T> extends StatefulWidget {
   /// [builder] provides the child widget to rendered.
-  final VxStateWidgetBuilder builder;
+  final VxStateWidgetBuilder<T> builder;
 
   /// Widget will rerender every time any of [mutations] executes.
   final Set<Type> mutations;
 
   /// Map of mutations and their corresponding callback
-  final Map<Type, ContextCallback> notifications;
+  final Map<Type, ContextCallback>? notifications;
 
   /// Creates widget to rerender child widgets when given
   /// [mutations] execute.
-  VxConsumer({
-    @required this.builder,
-    @required this.mutations,
-    @required this.notifications,
-  })  : assert(mutations != null),
-        assert(notifications != null);
+  const VxConsumer({
+    required this.builder,
+    required this.mutations,
+    this.notifications,
+  });
 
   @override
-  _VxConsumerState createState() => _VxConsumerState();
+  _VxConsumerState createState() => _VxConsumerState<T>();
 }
 
-class _VxConsumerState extends State<VxConsumer> {
-  StreamSubscription eventSub;
+class _VxConsumerState<T> extends State<VxConsumer<T>> {
+  StreamSubscription? eventSub;
 
   @override
   void initState() {
     super.initState();
-    final notifications = widget.notifications.keys.toSet();
-    final stream = VxState.events.where(
-      (e) => notifications.contains(e.runtimeType),
-    );
-    eventSub = stream.listen((e) {
-      widget.notifications[e.runtimeType]?.call(context, e);
-    });
+    if (widget.notifications != null) {
+      final notifications = widget.notifications!.keys.toSet();
+      final stream = VxState.events.where(
+        (e) => notifications.contains(e.runtimeType),
+      );
+      eventSub = stream.listen((e) {
+        widget.notifications![e.runtimeType]
+            ?.call(context, e, status: e.status);
+      });
+    }
   }
 
   @override
@@ -59,13 +61,14 @@ class _VxConsumerState extends State<VxConsumer> {
     return StreamBuilder<VxMutation>(
       stream: stream,
       builder: (context, mut) {
-        VxStatus status;
+        VxStatus? status;
         if (!mut.hasData || mut.connectionState == ConnectionState.waiting) {
           status = VxStatus.none;
         } else {
-          status = mut?.data?.status;
+          status = mut.data?.status;
         }
-        return widget.builder(context, status);
+        final T _store = VxState.store as T;
+        return widget.builder(context, _store, status);
       },
     );
   }