+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [2.3.0]
+
+- Store is not dynamic anymore.
+- Minor upgrades
+
+## [2.2.0]
+
+- VxConsumer improved with notifications as optional and also returns status.
+- Bug Fixes
+
+## [2.1.0]
+
+- VxBuilder & VxConsumer now returns store.
+- Store can be directly accessed as MyStore store = VxState.store
+- Status for all async operations
+- Bug Fixes
+
+## [2.0.1]
+
+- VxState.listen(to: ) is now VxState.watch(on: )
+
+## [2.0.0]
+
+- Null Safety Enabled
+
 ## [1.0.1]
 
 - Enhancements.